requirements.txt
unstructured
unstructured[pdf]
pytesseract
pdf2image
pdfplumber
gradio
uvicorn
langchain
langchain-community
langchain-openai
sentence-transformers
transformers
torch
torchvision
faiss-cpu
rank-bm25
numpy
openai
requests
python-dotenv
pillow
opencv-python
pathlib
tempfile
shutil
pickle
json
re
os
base64